moved test utility into dedicated file
