+++ packages/langium/test/grammar/lsp/grammar-code-actions.test.ts
@@ -4,20 +4,16 @@
  * terms of the MIT License, which is available in the project root.
  ******************************************************************************/
 
-import { AstNode, AsyncDisposable, EmptyFileSystem, LangiumDocument } from 'langium';
-import { ParseHelperOptions, ValidationResult, textDocumentParams, validationHelper } from 'langium/test';
+import { EmptyFileSystem } from 'langium';
 import { describe, expect, test } from 'vitest';
-import type { Diagnostic } from 'vscode-languageserver';
-import { CodeAction } from 'vscode-languageserver';
-import { TextDocument } from 'vscode-languageserver-textdocument';
-import { createLangiumGrammarServices } from '../../../src/grammar/langium-grammar-module.js';
 import { IssueCodes } from '../../../src/grammar/index.js';
-import { LangiumServices } from '../../../lib/lsp/lsp-services.js';
+import { createLangiumGrammarServices } from '../../../src/grammar/langium-grammar-module.js';
+import { testQuickFix } from '../../../src/test/langium-test.js';
 
 const services = createLangiumGrammarServices(EmptyFileSystem);
 const testQuickFixes = testQuickFix(services.grammar);
 
-// Some of these test data are exported, since they are reused for corresponding test cases 
+// Some of these test data are exported, since they are reused for corresponding test cases for the grammar validation itself
 
 export const textBeforeParserRuleCrossReferences = `
     grammar ParserRulesOnlyForCrossReferences
@@ -87,91 +83,3 @@ describe('Langium grammar quick-fixes for validations: Parser rules used only as
 
     // TODO test cases, where no quick-fix action is provided
 });
-
-export interface QuickFixResult<T extends AstNode = AstNode> extends AsyncDisposable {
-    /** the document containing the AST */
-    document: LangiumDocument<T>;
-    /** all diagnostics of the validation */
-    diagnosticsAll: Diagnostic[];
-    /** the relevant Diagnostic with the given diagnosticCode, it is expected that the given input has exactly one such diagnostic */
-    diagnosticRelevant: Diagnostic;
-    /** the CodeAction to fix the found relevant problem, it is possible, that there is no such code action */
-    action?: CodeAction;
-}
-/**
- * This is a helper function to easily test quick-fixes for validation problems.
- * @param services the Langium services for the language with quick fixes
- * @returns A function to easily test a single quick-fix on the given invalid 'input'.
- * This function expects, that 'input' contains exactly one validation problem with the given 'diagnosticCode'.
- * If 'outputAfterFix' is specified, this functions checks, that the diagnostic comes with a single quick-fix for this validation problem.
- * After applying this quick-fix, 'input' is transformed to 'outputAfterFix'.
- */
-export function testQuickFix<T extends AstNode = AstNode>(services: LangiumServices):
-        (input: string, diagnosticCode: string, outputAfterFix: string | undefined, options?: ParseHelperOptions) => Promise<QuickFixResult<T>> {
-    const validateHelper = validationHelper<T>(services);
-    return async (input, diagnosticCode, outputAfterFix, options) => {
-        // parse + validate
-        const validationBefore = await validateHelper(input, options);
-        const document = validationBefore.document;
-        const diagnosticsAll = document.diagnostics ?? [];
-        // use only the diagnostics with the given validation code
-        const diagnosticsRelevant = diagnosticsAll.filter(d => d.data && 'code' in d.data && d.data.code === diagnosticCode);
-        // expect exactly one validation with the given code
-        expect(diagnosticsRelevant.length).toBe(1);
-        const diagnosticRelevant = diagnosticsRelevant[0];
-
-        // check, that the quick-fixes are generated for the selected validation:
-        // prepare the action provider
-        const actionProvider = services.lsp.CodeActionProvider;
-        expect(actionProvider).toBeTruthy();
-        // request the actions for this diagnostic
-        const currentActions = await actionProvider!.getCodeActions(document, {
-            ...textDocumentParams(document),
-            range: diagnosticRelevant.range,
-            context: {
-                diagnostics: diagnosticsRelevant,
-                triggerKind: 1 // explicitly triggered by users (or extensions)
-            }
-        });
-
-        // evaluate the resulting actions
-        let action: CodeAction | undefined;
-        let validationAfter: ValidationResult | undefined;
-        if (outputAfterFix) {
-            // exactly one quick-fix is expected
-            expect(currentActions).toBeTruthy();
-            expect(Array.isArray(currentActions)).toBeTruthy();
-            expect(currentActions!).toHaveLength(1);
-            expect(CodeAction.is(currentActions![0])).toBeTruthy();
-            action = currentActions![0] as CodeAction;
-
-            // execute the found quick-fix
-            const edits = action.edit?.changes![document.textDocument.uri];
-            expect(edits).toBeTruthy();
-            const updatedText = TextDocument.applyEdits(document.textDocument, edits!);
-
-            // check the result after applying the quick-fix:
-            // 1st text is updated as expected
-            expect(updatedText).toBe(outputAfterFix);
-            // 2nd the validation diagnostic is gone after the fix
-            validationAfter = await validateHelper(updatedText, options);
-            const diagnosticsUpdated = validationAfter.diagnostics.filter(d => d.data && 'code' in d.data && d.data.code === diagnosticCode);
-            expect(diagnosticsUpdated).toHaveLength(0);
-        } else {
-            // no quick-fix is expected
-            expect(currentActions).toBeFalsy();
-        }
-
-        async function dispose(): Promise<void> {
-            validationBefore.dispose();
-            validationAfter?.dispose();
-        }
-        return {
-            document,
-            diagnosticsAll,
-            diagnosticRelevant: diagnosticRelevant,
-            action,
-            dispose: () => dispose()
-        };
-    };
-}
